Pulling UCX-specific functionality from cumlCommunicator into helper class
